+++ Data/Config.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using AutoMapPins.Common;
 using UnityEngine;
-
+// ReSharper disable ConvertToConstant.Global
 // ReSharper disable FieldCanBeMadeReadOnly.Global
 // ReSharper disable CollectionNeverUpdated.Global
 // ReSharper disable UnassignedField.Global
@@ -12,43 +12,41 @@ namespace AutoMapPins.Data;
 
 public abstract class Config
 {
-    public class Category
+    public class Category : Pin
     {
-        public bool CategoryActive;
-        public Dictionary<string, Pin> Pins = new();
-
-        public override string ToString()
-        {
-            return YamlFileStorage.Serializer.Serialize(this);
-        }
+        public Dictionary<string, Pin>? IndividualConfiguredObjects = null!;
+        public List<string>? CategoryConfiguredObjects = null!;
     }
-
-    public class Pin
+    
+    public class Pin : SerializedToStrings
     {
-        public string Name = null!;
-        public string IconName = Constants.NoConfig;
-        public PinColor IconColorRGBA = Constants.WhiteColor;
+        public string? Name = null!;
+        public string? IconName = null!;
+        public PinColor? IconColorRGBA = null!;
         public bool IsPermanent;
         public bool IsActive;
         public bool Groupable;
         public float GroupingDistance = Constants.DefaultGroupingDistance;
-
-        public override string ToString()
-        {
-            return YamlFileStorage.Serializer.Serialize(this);
-        }
     }
 
-    public class PinColor
+    public class PinColor : SerializedToStrings
     {
-        public float Red = 0f;
-        public float Green = 0f;
-        public float Blue = 0f;
+        public float Red = 1f;
+        public float Green = 1f;
+        public float Blue = 1f;
         public float Alpha = 1f;
 
         public Color FromConfig()
         {
             return new Color(Red, Green, Blue, Alpha);
+        }
+    }
+
+    public abstract class SerializedToStrings
+    {
+        public override string ToString()
+        {
+            return YamlFileStorage.Serializer.Serialize(this);
         }
     }
 }