+++ Data/Registry.cs
@@ -19,45 +19,89 @@ public abstract class Registry : HasLogger
         MissingConfigs.Clear();
         ConfiguredObjects.Clear();
         ConfiguredPins.Clear();
-        RegisterConfiguredObjects(newConfiguredCategories);
         ConfiguredPins = LoadActivePinConfigs(newConfiguredCategories);
+        ConfiguredObjects = GetConfiguredObjects(newConfiguredCategories);
         if (updatePins) Map.UpdatePins();
         Log.LogDebug("registry initialization successful");
     }
 
-    private static void RegisterConfiguredObjects(Dictionary<string, Config.Category> newConfiguredCategories)
-    {
-        List<string> objectNames = newConfiguredCategories
-            .SelectMany(categories => categories.Value.Pins)
-            .Select(pin => pin.Key).ToList();
-
-        ConfiguredObjects = objectNames.Distinct().ToHashSet();
-
-        Dictionary<string, int> duplicatedConfigs = objectNames
-            .GroupBy(name => name)
-            .ToDictionary(group => group.Key, group => group.Count())
-            .Where(kv => kv.Value > 1)
-            .ToDictionary(kv => kv.Key, kv => kv.Value);
-
-        if (duplicatedConfigs.Count <= 0) return;
-        Log.LogWarning("found duplicates in configured game objects:");
-        foreach (var config in duplicatedConfigs)
-            AutoMapPinsPlugin.Log.LogWarning($"name '{config.Key}' has '{config.Value}' entries in config");
-
-        Log.LogWarning(
-            "duplicates are not allowed, only first entry will be used, please check and fix the config files");
-    }
 
     private static Dictionary<string, Config.Pin> LoadActivePinConfigs(
         Dictionary<string, Config.Category> newConfiguredCategories)
     {
-        Dictionary<string, Config.Pin> activePinConfigs = newConfiguredCategories
-            .Where(category => category.Value.CategoryActive)
-            .SelectMany(x => x.Value.Pins)
-            .Where(pin => pin.Value.IsActive)
-            .ToDictionary(kv => kv.Key, kv => kv.Value);
+        var activePinConfigs = new Dictionary<string, Config.Pin>();
+        foreach (KeyValuePair<string, Config.Category> category in newConfiguredCategories)
+        {
+            if (!category.Value.IsActive) continue;
+            if (category.Value.IndividualConfiguredObjects != null)
+                foreach (KeyValuePair<string, Config.Pin> configuredObject in
+                         category.Value.IndividualConfiguredObjects)
+                {
+                    if (configuredObject.Value is not { IsActive: true }) continue;
+                    var objectFixedByCategory = configuredObject.Value;
+                    objectFixedByCategory.Name ??= category.Value.Name;
+                    objectFixedByCategory.IconName ??= category.Value.IconName;
+                    objectFixedByCategory.IconColorRGBA ??= category.Value.IconColorRGBA;
+                    activePinConfigs.Add(configuredObject.Key, objectFixedByCategory);
+                }
+
+            if (category.Value.CategoryConfiguredObjects != null)
+                foreach (string configuredObject in category.Value.CategoryConfiguredObjects)
+                {
+                    if (!activePinConfigs.ContainsKey(configuredObject))
+                        activePinConfigs.Add(configuredObject, new Config.Pin
+                        {
+                            Name = category.Value.Name,
+                            IsActive = category.Value.IsActive,
+                            IsPermanent = category.Value.IsPermanent,
+                            Groupable = category.Value.Groupable,
+                            GroupingDistance = category.Value.GroupingDistance,
+                            IconName = category.Value.IconName,
+                            IconColorRGBA = category.Value.IconColorRGBA
+                        });
+                    else
+                        Log.LogWarning(
+                            $"tried to add a by category '{category.Key}' configured object '{configuredObject}' " +
+                            $"that does already exist in individual configured objects list - skipping object");
+                }
+        }
+
         Log.LogInfo($"loaded '{activePinConfigs.Count}' configs for active pins");
         return activePinConfigs;
+    }
+
+    private static HashSet<string> GetConfiguredObjects(Dictionary<string, Config.Category> newConfiguredCategories)
+    {
+        List<string> categoryConfiguredTotal = newConfiguredCategories
+            .Where(cat => cat.Value.CategoryConfiguredObjects != null)
+            .SelectMany(cat => cat.Value.CategoryConfiguredObjects).ToList();
+
+        List<string> individualConfiguredTotal = new List<string>();
+        foreach (var category in newConfiguredCategories)
+        {
+            if (category.Value.IndividualConfiguredObjects == null) continue;
+            foreach (var configuredObject in category.Value.IndividualConfiguredObjects)
+                individualConfiguredTotal.Add(
+                    (configuredObject.Value.Name ?? category.Value.Name) ?? Constants.NoConfig);
+        }
+
+        var configuredTotal = categoryConfiguredTotal.Concat(individualConfiguredTotal).ToList();
+        var duplicates = configuredTotal.GroupBy(x => x)
+            .ToDictionary(group => group.Key, group => group.Count())
+            .Where(kv => kv.Value > 1)
+            .ToDictionary(kv => kv.Key, kv => kv.Value);
+        if (duplicates.Count > 0)
+        {
+            Log.LogWarning($"found '{duplicates.Count}' duplicates in config:");
+            foreach (var duplicate in duplicates)
+            {
+                Log.LogWarning($"object name '{duplicate.Key}' found '{duplicate.Value}' times");
+            }
+        }
+
+        var result = configuredTotal.Distinct().ToHashSet();
+        Log.LogInfo($"loaded '{result.Count}' configured objects in total (also inactive ones)");
+        return result;
     }
 
     internal static void AddMissingConfig(string internalName)